refactor(문제6): 검증 기능 객체 분리
